Haena_P_최댓값과 최솟값
조해나 프로그래머스 최댓값과 최솟값
배열을 일정한 기준으로 나누는 메소드를 몰라 구글링을 통해 split 메소드를 공부하여 사용하였고, 이후에 String 배열을 int 배열로 바꾸고, int 배열에서 최댓값과 최솟값을 구하는 문제는 이전에 풀어본 적이 있어 수월하게 풀 수 있었습니다. 답은 String 으로 나와야하기 때문에 최댓값 최솟값을 구한 후 다시 String 으로 변환하여 풀이하였습니다.
