--- src/Test06/Test0612_P/Hyena.java
@@ -1,27 +1,0 @@
-// 최댓값과 최솟값
-
-package Test06.Test0612_P;
-
-public class Hyena {
-    // 예시를 이용하여 solution 메서드를 잘 구현했는지 확인하는 main 메서드
-    public static void main( String[] args ) {
-        Hyena hn = new Hyena();
-
-        // 프로그래머스 문제에서 나오는 예제들
-        String t1 = "1 2 3 4";
-        String t2 = "-1 -2 -3 -4";
-        String t3 = "-1 -1";
-
-        System.out.println( hn.solution( t1 ) );
-        System.out.println( hn.solution( t2 ) );
-        System.out.println( hn.solution( t3 ) );
-
-    }
-
-    // solution 메서드
-    public String solution( String s ) {
-        String answer = "";
-        // 문제 푸는 부분
-        return answer;
-    }
-}
